move & refactor pdo_select_type.php move & refactor functions.php
