+++ PULL_OUT_TO_SERVER/README_DRM.txt
@@ -1,0 +1,3 @@
+Dr M,
+This file should be extracted totally from the directories of the site, set on the server as execute only in its own directory.
+~Dave Babler
