+++ Scripts/DB/pdo_select_type.php
