+++ insert.php
@@ -1,7 +1,7 @@
 <?php
 session_start();
 include('db.php');
-include('function.php');
+include('/home/dbabler/dbabler.yaacotu.com/FED_2020/Scripts/DB/function.php');
 if(isset($_POST["operation"]))
 {
 	if($_POST["operation"] == "Add")
